fix compile under windows
svn path=/trunk/KDE/kdelibs/; revision=690236

+++ solid/solid/backends/hal/halstorageaccess.cpp
@@ -150,7 +150,7 @@ bool StorageAccess::requestPassphrase()
     QWidget *activeWindow = QApplication::activeWindow();
     uint wId = 0;
     if (activeWindow!=0) {
-        wId = activeWindow->winId();
+        wId = (uint)activeWindow->winId();
     }
 
     QString appId = QCoreApplication::applicationName();
